+++ src/cylinder_bank_object.cpp
@@ -13,13 +13,13 @@ CylinderBankObject::~CylinderBankObject() {
 void CylinderBankObject::generateGeometry() {
     GeometryGenerator *gen = m_app->getGeometryGenerator();
 
-    const float lineWidth = m_bank->m_bore * 0.1f;
-    const float dx = -m_bank->m_dy * (m_bank->m_bore / 2 + lineWidth / 2);
-    const float dy = m_bank->m_dx * (m_bank->m_bore / 2 + lineWidth / 2);
-    const float top_x = m_bank->m_dx * m_bank->m_deckHeight;
-    const float top_y = m_bank->m_dy * m_bank->m_deckHeight;
-    const float bottom_x = m_bank->m_dx * (0.4f * m_bank->m_deckHeight);
-    const float bottom_y = m_bank->m_dy * (0.4f * m_bank->m_deckHeight);
+    const float lineWidth = (float)(m_bank->m_bore * 0.1);
+    const float dx = -(float)(m_bank->m_dy * (m_bank->m_bore / 2 + lineWidth / 2));
+    const float dy = (float)(m_bank->m_dx * (m_bank->m_bore / 2 + lineWidth / 2));
+    const float top_x = (float)(m_bank->m_dx * m_bank->m_deckHeight);
+    const float top_y = (float)(m_bank->m_dy * m_bank->m_deckHeight);
+    const float bottom_x = (float)(m_bank->m_dx * (0.4 * m_bank->m_deckHeight));
+    const float bottom_y = (float)(m_bank->m_dy * (0.4 * m_bank->m_deckHeight));
 
     GeometryGenerator::Line2dParameters params;
     params.lineWidth = lineWidth;
